V2/web/requirements.txt
# AIWA Web Application Dependencies
# 用于 Docker 构建的完整依赖列表

# ========== Web Framework ==========
flask>=2.3.0
flask-cors>=4.0.0

# ========== LLM API ==========
openai>=1.0.0

# ========== Browser Automation ==========
playwright>=1.40.0

# ========== HTTP Clients ==========
requests>=2.28.0
httpx>=0.25.0
aiohttp>=3.9.0

# ========== Document Processing ==========
python-docx>=1.0.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
pandas>=2.0.0

# ========== Environment Variables ==========
python-dotenv>=1.0.0

# ========== Database & Cache ==========
redis>=5.0.0

# ========== Async Support ==========
gevent>=23.9.0
eventlet>=0.33.0

# ========== Production Server ==========
gunicorn>=21.0.0

# ========== Utilities ==========
markdown>=3.5.0
jinja2>=3.1.0
